# Profile native compiled wasm