--- a/core/executor/src/lib.rs
+++ b/core/executor/src/lib.rs
@@ -29,6 +29,9 @@
 #![recursion_limit="128"]
 
 #[macro_use]
+extern crate lazy_static;
+
+#[macro_use]
 mod wasm_utils;
 mod wasm_executor;
 #[macro_use]