--- a/core/executor/src/wasm_executor.rs
+++ b/core/executor/src/wasm_executor.rs
@@ -19,6 +19,7 @@
 use std::collections::HashMap;
 use std::mem::ManuallyDrop;
 use std::sync::{Mutex, MutexGuard};
+use std::time::Instant;
 use tiny_keccak;
 use secp256k1;
 
@@ -1113,6 +1114,8 @@ impl WasmExecutor {
 		method: &str,
 		data: &[u8],
 	) -> Result<Vec<u8>> {
+		let start = Instant::now();
+
 		// extract a reference to a linear memory, optional reference to a table
 		// and then initialize FunctionExecutor.
 		let memory = Self::get_mem_instance(module_instance)?;
@@ -1172,6 +1175,8 @@ impl WasmExecutor {
 		memory.with_direct_access_mut(|buf| buf.resize(used_mem.0, 0));
 		*/
 
+		let duration = start.elapsed();
+		eprintln!("duration for {:?}: {:?}µs", method, duration.as_micros());
 		result
 	}