… core/test-runtime/wasm/target/wasm32-unknown-unknown/release/substrate_test_runtime.compact.wasm
… libnode_runtime.so
… node/runtime/wasm/target/wasm32-unknown-unknown/release/node_runtime.compact.wasm
… wasm-rt.h