This addresses #1519 and this PR is not supposed to be merged, it's just here for discussion and can be closed afterwards.

The idea is to gain insight into how much optimization room we have for calls into wasm. To get an idea of the "ground truth" I use [wasm2c](https://github.com/WebAssembly/wabt/tree/master/wasm2c) to convert our `node_runtime.wasm` to C, which I then compile as a shared library, which I link to the substrate executable.

I then timed invocations of `call_in_wasm_module` in a very simple manner and ran `cargo run -- --dev` for 30 blocks. I did the same for master (here is the timed master branch: https://github.com/paritytech/substrate/tree/cmichi-time-wasm-calls, the same timing patch was applied + method invocations are forced into wasm). No interactions took place, only empty blocks were produced (though there already happens a lot of stuff, even for empty blocks).

This is only a very rough benchmark! I'm totally aware of that. For example, system interruptions by the OS scheduler are not accounted for and could skew the results. But: a better benchmark ‒ e.g. benchmarking  `--import-blocks` of a chain dump and monitoring via `perf-stat(1)` ‒ would require a lot more work. This is because substrate handles multiple wasmi `ModuleInstances` and this branch uses shared memory with the so, which basically means only one "instance" can be run at a time.

Either way, I ran the measurements on an otherwise idling system and I think the results are sufficient to give a rough idea how much headroom there is for the different methods.

```
# Core_execute_block (average over 31 measurements)
from_so:       7064 µs
 master:      58857 µs

# Core_version (average over 7 measurements)
from_so:        136 µs
 master:       5542 µs

# Core_authorities (average over 31 measurements)
from_so:        491 µs
 master:      15285 µs

# Core_initialise_block (average over 95 measurements)
from_so:       7446 µs
 master:      44112 µs

# BlockBuilder_apply_extrinsic (average over 31 measurements)
from_so:       1290 µs
 master:      14221 µs

# BlockBuilder_finalise_block (average over 31 measurements)
from_so:       5326 µs
 master:      11596 µs

# BlockBuilder_inherent_extrinsics (average over 31 measurements)
from_so:        464 µs
 master:       7154 µs

# AuraApi_slot_duration (average over 1 measurements)
from_so:        283 µs
 master:       5684 µs

# GrandpaApi_grandpa_authorities (average over 1 measurements)
from_so:        417 µs
 master:       5476 µs

# GrandpaApi_grandpa_pending_change (average over 31 measurements)
from_so:        127 µs
 master:       4449 µs
```