@Vurich 2019-02-21
This is really fantastic - it's a great goal to aim towards if nothing else. One way to mitigate the cost of functions that will be called many hundreds of thousands of times like the runtime functions is to patch the calls directly instead of using the relocation table, but I don't know if that's necessary since for the runtime we expect users wanting the most speed to use native code directly anyway. A couple questions:

- What's going on with `AuraApi_slot_duration`?
- I don't think I quite follow the testing methodology - you're trying to have it exit immediately by processing no transactions so you only test the calling cost? That doesn't take into account inefficiencies like stack overflow checking or even just poor compilation of the Wasm itself even in the small amount of code that gets executed.