@cmichi 2019-02-21
The initial version of the numbers in the first post was gathered without `--release`, I have updated them with everything re-executed in release mode.

@Vurich 
> What's going on with `AuraApi_slot_duration`?

No idea, for `AuraApi_slot_duration` and `GrandpaApi_grandpa_pending_change` wasmi strangely performs better than the shared object. I'll need to look at the implementation to see if something comes up.

> I don't think I quite follow the testing methodology - you're trying to have it exit immediately by processing no transactions so you only test the calling cost? That doesn't take into account inefficiencies like stack overflow checking or even just poor compilation of the Wasm itself even in the small amount of code that gets executed.

Yeah, I totally agree, processing transactions would of course provide better insights. This limitation was not done intentionally though, it's just the lowest hanging fruit to test. Getting it to work for a more complex chain would require a more extensive implementation