@rphmeier 2019-02-21
> What's going on with `AuraApi_slot_duration`

It's only called at genesis, FWIW