@cmichi 2019-02-22
Found the issue with `AuraApi_slot_duration` and `GrandpaApi_grandpa_pending_change` ‒ I moved the timer in front of the immediate invocation of the method in wasmi/wasm2c. In the prior version I had timed the `call_in_wasm_module` method, which did a bit more then just invoke the method externally.

I updated the first post if the re-run metrics.